refactor: git.rs to a git folder
